requirements.txt
# FastAPI and web server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data processing
pandas>=2.1.4
numpy>=1.24.3

# Google APIs
gspread>=5.12.4
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# HTTP requests
requests>=2.31.0
urllib3>=2.1.0

# Additional dependencies for production
python-multipart>=0.0.6
